more refactoring post merge
