+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/domain/State.java
@@ -33,8 +33,7 @@ public enum State {
     DRAFT,
 
     @CCD(
-        label = "Withdrawn",
-        access = {CaseworkerAccess.class}
+        label = "Withdrawn"
     )
     WITHDRAWN;
