+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/entity/AddressEntity.java
@@ -24,7 +24,7 @@ import static jakarta.persistence.FetchType.LAZY;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Address {
+public class AddressEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
