+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/event/CaseworkerUpdateApplication.java
@@ -7,13 +7,13 @@ import uk.gov.hmcts.ccd.sdk.api.CCDConfig;
 import uk.gov.hmcts.ccd.sdk.api.ConfigBuilder;
 import uk.gov.hmcts.ccd.sdk.api.EventPayload;
 import uk.gov.hmcts.ccd.sdk.api.Permission;
+import uk.gov.hmcts.reform.pcs.ccd.accesscontrol.UserRole;
 import uk.gov.hmcts.reform.pcs.ccd.domain.PCSCase;
 import uk.gov.hmcts.reform.pcs.ccd.domain.State;
-import uk.gov.hmcts.reform.pcs.ccd.accesscontrol.UserRole;
-import uk.gov.hmcts.reform.pcs.ccd.service.PcsCaseService;
+import uk.gov.hmcts.reform.pcs.ccd.service.PCSCaseService;
 
-import static uk.gov.hmcts.reform.pcs.ccd.domain.State.CASE_ISSUED;
 import static uk.gov.hmcts.reform.pcs.ccd.accesscontrol.UserRole.PCS_CASE_WORKER;
+import static uk.gov.hmcts.reform.pcs.ccd.domain.State.CASE_ISSUED;
 import static uk.gov.hmcts.reform.pcs.ccd.event.EventId.caseworkerUpdateApplication;
 
 @Component
@@ -21,7 +21,7 @@ import static uk.gov.hmcts.reform.pcs.ccd.event.EventId.caseworkerUpdateApplicat
 @AllArgsConstructor
 public class CaseworkerUpdateApplication implements CCDConfig<PCSCase, State, UserRole> {
 
-    private final PcsCaseService pcsCaseService;
+    private final PCSCaseService pcsCaseService;
 
     @Override
     public void configure(final ConfigBuilder<PCSCase, State, UserRole> configBuilder) {
