+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/service/PcsCaseService.java
@@ -1,30 +1,69 @@
 package uk.gov.hmcts.reform.pcs.ccd.service;
 
-import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.ccd.sdk.type.AddressUK;
+import uk.gov.hmcts.ccd.sdk.type.ListValue;
+import uk.gov.hmcts.reform.pcs.ccd.domain.GACase;
 import uk.gov.hmcts.reform.pcs.ccd.domain.PCSCase;
+import uk.gov.hmcts.reform.pcs.ccd.entity.AddressEntity;
+import uk.gov.hmcts.reform.pcs.ccd.entity.ClaimantInfo;
+import uk.gov.hmcts.reform.pcs.ccd.entity.GACaseEntity;
+import uk.gov.hmcts.reform.pcs.ccd.entity.PCSCaseEntity;
 import uk.gov.hmcts.reform.pcs.ccd.repository.PCSCaseRepository;
 import uk.gov.hmcts.reform.pcs.exception.CaseNotFoundException;
 
-@Service
-@AllArgsConstructor
-public class PcsCaseService {
+import java.util.stream.Collectors;
 
-    private final PCSCaseRepository pcsCaseRepository;
+@Service
+
+public class PCSCaseService {
+
+
     private final ModelMapper modelMapper;
+    private final PCSCaseRepository pcsCaseRepository;
+
+    public PCSCaseService(ModelMapper modelMapper, PCSCaseRepository pcsCaseRepository) {
+        this.modelMapper = modelMapper;
+        this.pcsCaseRepository = pcsCaseRepository;
+    }
+
+    public PCSCase convertToPCSCase(PCSCaseEntity pcs) {
+        PCSCase.PCSCaseBuilder builder = PCSCase.builder()
+            .caseReference(pcs.getCaseReference());
+        if (pcs.getPropertyAddress() != null) {
+            builder.propertyAddress(convertAddress(pcs.getPropertyAddress()));
+        }
+        if (pcs.getClaimantInfo() != null) {
+            builder.applicantForename(pcs.getClaimantInfo().getForename());
+            builder.applicantSurname(pcs.getClaimantInfo().getSurname());
+        }
+        if (pcs.getGeneralApplications() != null) {
+            builder.generalApplications(
+                pcs.getGeneralApplications().stream()
+                    .map(ga -> {
+                        GACase dto = convertGenApplication(ga);
+                        return ListValue.<GACase>builder()
+                            .id(ga.getId().toString())
+                            .value(dto)
+                            .build();
+                    })
+                    .collect(Collectors.toList())
+            );
+        }
+        return builder.build();
+    }
 
     public void createCase(long caseReference, PCSCase pcsCase) {
         AddressUK applicantAddress = pcsCase.getPropertyAddress();
 
         AddressEntity addressEntity = applicantAddress != null
             ? modelMapper.map(applicantAddress, AddressEntity.class) : null;
-
+        ClaimantInfo claimantInfo = createClaimantInfoEntity(pcsCase);
         PCSCaseEntity pcsCaseEntity = new PCSCaseEntity();
         pcsCaseEntity.setCaseReference(caseReference);
-        pcsCaseEntity.setApplicantForename(pcsCase.getApplicantForename());
-        pcsCaseEntity.setApplicantSurname(pcsCase.getApplicantSurname());
+        pcsCaseEntity.setClaimantInfo(claimantInfo);
+        ;
         pcsCaseEntity.setPropertyAddress(addressEntity);
 
         pcsCaseRepository.save(pcsCaseEntity);
@@ -34,12 +73,16 @@ public class PcsCaseService {
         PCSCaseEntity pcsCaseEntity = pcsCaseRepository.findByCaseReference(caseReference)
             .orElseThrow(() -> new CaseNotFoundException(caseReference));
 
+        ClaimantInfo claimantInfo = new ClaimantInfo();
+
+
         if (pcsCase.getApplicantForename() != null) {
-            pcsCaseEntity.setApplicantForename(pcsCase.getApplicantForename());
+            claimantInfo.setForename(pcsCase.getApplicantForename());
         }
 
         if (pcsCase.getApplicantSurname() != null) {
-            pcsCaseEntity.setApplicantSurname(pcsCase.getApplicantSurname());
+            claimantInfo.setForename(pcsCase.getApplicantSurname());
+            pcsCaseEntity.setClaimantInfo(claimantInfo);
         }
 
         if (pcsCase.getPropertyAddress() != null) {
@@ -50,4 +93,38 @@ public class PcsCaseService {
         pcsCaseRepository.save(pcsCaseEntity);
     }
 
+    public PCSCaseEntity convertToPCSEntity(PCSCase pcsCase) {
+        return modelMapper.map(pcsCase, PCSCaseEntity.class);
+    }
+
+    private AddressUK convertAddress(AddressEntity address) {
+        return AddressUK.builder()
+            .addressLine1(address.getAddressLine1())
+            .addressLine2(address.getAddressLine2())
+            .postTown(address.getPostTown())
+            .county(address.getCounty())
+            .postCode(address.getPostCode())
+            .country(address.getCountry())
+            .build();
+    }
+
+    private GACase convertGenApplication(GACaseEntity ga) {
+        return GACase.builder()
+            .caseReference(ga.getCaseReference())
+            .adjustment(ga.getAdjustment())
+            .additionalInformation(ga.getAdditionalInformation())
+            .status(ga.getStatus())
+            .build();
+    }
+
+    public PCSCaseEntity findPCSCase(Long caseReference) {
+        return pcsCaseRepository.findByCaseReference(caseReference)
+            .orElseThrow(() -> new IllegalStateException("PCS case not found"));
+    }
+
+    public ClaimantInfo createClaimantInfoEntity(PCSCase pcsCase) {
+        ClaimantInfo claimantInfo = ClaimantInfo.builder().forename(pcsCase.getApplicantForename())
+            .surname(pcsCase.getApplicantSurname()).build();
+        return claimantInfo;
+    }
 }
