+++ src/test/java/uk/gov/hmcts/reform/pcs/ccd/event/CitizenUpdateApplicationTest.java
@@ -14,7 +14,7 @@ import uk.gov.hmcts.reform.pcs.ccd.ShowConditions;
 import uk.gov.hmcts.reform.pcs.ccd.accesscontrol.UserRole;
 import uk.gov.hmcts.reform.pcs.ccd.domain.PCSCase;
 import uk.gov.hmcts.reform.pcs.ccd.domain.State;
-import uk.gov.hmcts.reform.pcs.ccd.service.PcsCaseService;
+import uk.gov.hmcts.reform.pcs.ccd.service.PCSCaseService;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
@@ -28,7 +28,7 @@ import static uk.gov.hmcts.reform.pcs.ccd.domain.State.AWAITING_SUBMISSION_TO_HM
 class CitizenUpdateApplicationTest extends BaseEventTest {
 
     @Mock
-    private PcsCaseService pcsCaseService;
+    private PCSCaseService pcsCaseService;
 
     private Event<PCSCase, UserRole, State> configuredEvent;
